rename component-detail into codelist-detail
