+++ packages/codelists/src/components/codelist-detail/view-container.js
