+++ packages/codelists/src/components/codelist-detail/view.js
