+++ packages/codelists/src/index.js
@@ -3,4 +3,4 @@ export { default as D, D1, D2 } from './i18n/build-dictionary';
 export { default as Menu } from './menu';
 export { default as Home } from './components/home/home';
 export { default as SearchFormList } from './components/search/search';
-export { default as CodelistComponentView } from './components/component-detail/view-container';
+export { default as CodelistComponentView } from './components/codelist-detail/view-container';
